feat(CtkcFCld) Transaction

+++ database/migrations/2024_01_16_002932_create_employees_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('agama');
             $table->enum('status nikah', ['belum nikah', 'nikah']);
             $table->text('alamat');
-            $table->unsignedBigInteger('golongan_id');
+            $table->integer('golongan_id');
             $table->string('image');
             $table->timestamps();
         });
